requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
ShopifyAPI==12.6.0
django-autoslug==1.9.9
django-filter==24.3
gunicorn==23.0.0
pillow==10.4.0